Problem: /newproblem users are not guided enough when reporting problems
Solution: used mdg-validated method to add simple schema limit on header, and tried to errors from validation customizable (but they merely work with alerts) #649
