+++ .meteor/packages
@@ -55,3 +55,4 @@ meteorhacks:subs-manager
 practicalmeteor:mocha
 natestrauser:x-editable-bootstrap
 mrt:combodate
+mdg:validated-method
