+++ imports/api/problems/server/methods.js
@@ -7,71 +7,6 @@ import { UserData, Wallet } from '../../indexDB'
 import { check } from 'meteor/check'
 
 Meteor.methods({
-	newProblem: (type, header, text, images, bounty) => {
-		check(type, String)
-		check(header, String)
-		check(text, String)
-		check(images, [String])
-		check(bounty, Number)
-
-		if (~['bug', 'feature', 'question'].indexOf(type)) {
-			if (Meteor.userId()) {
-				if (bounty > 0) { // check if the user can finance the bounty
-					let user = UserData.findOne({
-						_id: Meteor.userId()
-					})
-
-					if (user.balance < bounty) {
-						throw new Meteor.Error('Error.', 'Insufficient funds.')
-					}
-				}
-
-				Problems.insert({
-					type: type,
-					header: header,
-					text: text,
-					images: images,
-					createdBy: Meteor.userId(),
-					date: new Date().getTime(),
-					credit: [{
-						userId: Meteor.userId(),
-						bounty: bounty
-					}],
-					open: true,
-					solved: false,
-					taken: {},
-					locked: false,
-					cancelled: false,
-					votes: [],
-					score: 0
-				})
-
-				if (bounty > 0) { // take the bounty from user's wallet
-					UserData.upsert({
-						_id: Meteor.userId()
-					}, {
-						$inc: {
-							balance: -bounty
-						}
-					})
-
-				    Wallet.insert({
-				    	time: new Date().getTime(),
-				    	owner: Meteor.userId(),
-				    	type: 'transaction',
-				      	from: 'Blockrazor',
-				      	message: `${bounty} KZR has been reserved from your account for funding a problem.`,
-				      	amount: -bounty,
-				     	read: false
-				    })
-				}
-			} else {
-				throw new Meteor.Error('Error.', 'You have to be logged in.')
-			}
-		} else {
-			throw new Meteor.Error('Error.', 'Invalid type.')
-		}
-	},
 	takeProblem: (problemId, additional) => {
 		check(problemId, String)
 		check(additional, Object)
