+++ package.json
@@ -23,6 +23,7 @@
     "mime-types": "^2.1.17",
     "moment": "^2.20.1",
     "scrollmagic": "^2.0.5",
+    "simpl-schema": "^1.4.2",
     "sweetalert": "^2.1.0",
     "underscore": "^1.8.3",
     "xml2js": "^0.4.19"
